started implemeneting ekf, clean up repo

+++ src/scene/frontier_projection.c
@@ -5,7 +5,7 @@
 ColumnSummary *create_column_summaries(const OccupancyMap *occupancy_grid_3d,
                                        int rover_height_cells) {
     int column_count = occupancy_grid_3d->width * occupancy_grid_3d->depth;
-    ColumnSummary *column_summaries = malloc((size_t)column_count * sizeof(ColumnSummary));
+    ColumnSummary *column_summaries = malloc((int)column_count * sizeof(ColumnSummary));
     if (!column_summaries) {
         return NULL;
     }
@@ -23,6 +23,7 @@ void apply_updates_to_projected_map(ColumnSummary *column_summaries,
                                     const OccupancyMap *occupancy_grid_3d,
                                     OccupancyMap *occupancy_grid_2d,
                                     int rover_height_cells) {
+
     update_column_summaries(column_summaries, (VoxelUpdate *)updates, count, occupancy_grid_3d);
 
     for (int i = 0; i < count; i++) {
@@ -34,11 +35,12 @@ void apply_updates_to_projected_map(ColumnSummary *column_summaries,
         if (y >= rover_height_cells) {
             continue;
         }
-
+        // TODO: define +1.0 -0.2 more clearly
         ColumnSummary *col = &(column_summaries[x * occupancy_grid_3d->depth + z]);
         if (col->is_blocking_count > 0) {
             occupancy_grid_2d->data[z * occupancy_grid_2d->width + x] = OCCUPIED_THRESHOLD + 1.0f;
-        } else {
+        }
+        else {
             occupancy_grid_2d->data[z * occupancy_grid_2d->width + x] = FREE_THRESHOLD - 0.2f;
         }
     }
